Fixed handshake bug, added file mapping
Had a bug in the handshake where I was encoding pstrlen as 68 instead of 19
which was why I couldn't see any messages coming back and forth.

Also added the file mapping to keep trach of which pieces of which files we
have thus far downloaded.

Changed some of the formatting to be more pep8 compliant as well.